fix(npd): Add version now works — untyped $4 in audit jsonb_build_object threw, swallowed silently
Root cause (found via LIVE Postgres logs, not static analysis): create-version's audit-log INSERT passed an untyped bind param to jsonb_build_object('versionNumber', $4). jsonb_build_object args are "any", so Postgres failed at PREPARE time with 'could not determine data type of parameter $4', throwing the whole action -> caught as persistence_failed -> the createVersionAdapter stripped it to {ok:false} -> the editor swallowed it with no feedback. Net effect: NO formulation version was EVER created (every formulation stuck at v1) and the version selector stayed disabled (<2 versions) — exactly the owner's 'Add version doesn't add / can't select any version' report.

- create-version.ts: cast $4::int in the audit jsonb_build_object (PREPARE-verified clean on live DB).

- formulation-editor.tsx onCreateVersion: surface the error on ok:false (was silently swallowed), mirroring the createDraft pattern.

Swept the whole codebase for the same jsonb_build_object($N) untyped-param class: this was the ONLY real instance. pause-resume-wo.ts $7 looked similar but is pinned to text by a sibling ext_jsonb->>'...' = $7 usage (PREPARE-verified clean) — no false fix shipped.

Why earlier checks missed it: a service-role DO-block sim used PL/pgSQL literals (typed), so it could not reproduce an untyped-bind-param prepare error; only the live app_user path + DB logs exposed it.

+++ apps/web/app/(npd)/pipeline/[projectId]/formulation/_actions/create-version.ts
@@ -124,7 +124,11 @@ export async function createFormulationVersion(input: {
            $1::uuid,
            $2::uuid,
            'formulation.version_created',
-           jsonb_build_object('sourceVersionId', $3::uuid::text, 'versionNumber', $4),
+           -- $4 MUST be cast: jsonb_build_object args are "any", so an untyped
+           -- bind param fails prepare with "could not determine data type of
+           -- parameter $4" — which threw the whole action (→ persistence_failed,
+           -- silently swallowed by the UI) so NO version was ever created.
+           jsonb_build_object('sourceVersionId', $3::uuid::text, 'versionNumber', $4::int),
            $5::uuid
          )`,
         [row.id, created.id, sourceVersionId, created.version_number, ctx.userId],
